requirements.txt
Pillow==9.2.0
vietocr==0.3.9
torch==1.12.1
torchvision==0.13.1
python-multipart==0.0.5
paddleocr==2.5.0.3
paddlepaddle==2.5.1
numpy
PyYAML==6.0.1
ultralytics